mlp - added a "setInternalParameter" method to the top virtual Packet class.
Derived-classes can use this to set internal parameters in a class that
derives from Packet. This was added to overcome the issues with packet_fpga
and packet_fpga_short that need to be told what their number of samples is.

We can now replace the dreaded constructs such as

  _packet = dynamic_cast<Packet_hbd_fpgashort *>
      ( _event->getPacket(PROTOTYPE3_FEM::PACKET_ID));
  _packet->setNumSamples(PROTOTYPE3_FEM::NSAMPLES);

with

 _event->getPacket(PROTOTYPE3_FEM::PACKET_ID)->setInternalParameter(PROTOTYPE3_FEM::NSAMPLES);

without the need for an unsafe cast.

The 2 HBD classes are currently the only ones that override that call. Let's keep
it that way.

This is not meant to be a free-for-all to add the need for such a thing.
The philosophy was and is that the packet payload must have all the
required information so it can be decoded autonomously. This is merely
a stop-gap measure since the hbd classes violate this requirement.