mlp an important interface change:
the 20+ years old getEvtSequence etc calls are now declared const.

There is a new Event API getOriginBuffer() that allows us to retrive the
buffer number where this event came from. In particular in the online
monitoring that helps for debugging. Not all events are "born" from a
buffer, then this is 0.